fix priority bug and refine code

+++ src/command.h
@@ -2,12 +2,14 @@
 #ifndef COMMAND_INCLUDED
 #define COMMAND_INCLUDED
 
+#include <cstdint>
+
 #include <queue>
 #include <memory>
 #include <utility>
 
 template<typename T>
-int priority(const T& t) { return 0; }
+int64_t priority(const T& t) { return 0; }
 
 template<typename T>
 void execute(T& t) {}
@@ -18,7 +20,7 @@ public:
 	template<typename T>
 	command_object_t(T t) : self_(std::make_shared<model_t<T>>(std::move(t))) {}
 
-	friend int priority(const command_object_t& o);
+	friend int64_t priority(const command_object_t& o);
 	friend void execute(command_object_t& o);
 
 private:
@@ -26,7 +28,7 @@ private:
 	{
 		virtual ~concept_t() {}
 
-		virtual int priority_() const = 0;
+		virtual int64_t priority_() const = 0;
 		virtual void execute_() = 0;
 	};
 
@@ -35,7 +37,7 @@ private:
 	{
 		model_t(T t) : data_(std::move(t)) {}
 
-		int priority_() const { return priority(data_); }
+		int64_t priority_() const { return priority(data_); }
 		void execute_() { execute(data_); }
 
 		T data_;
@@ -44,7 +46,7 @@ private:
 	std::shared_ptr<concept_t> self_;
 };
 
-inline int priority(const command_object_t& o) { return o.self_->priority_(); }
+inline int64_t priority(const command_object_t& o) { return o.self_->priority_(); }
 inline void execute(command_object_t& o) { o.self_->execute_(); }
 
 inline bool operator<(const command_object_t& lhs, const command_object_t& rhs)
